I have finished my refactoring. More details in long message
I have refactored it to put the classes and main code in different files

I have merged Game and Tictactoe. Because we aren't implementing
multiple games, we do not need a base game class to be extended by a
class for TicTacToe.

I have removed the board history and ability to handle boards of sizes
other than 3x3 from the spec, as they are not truly needed and you
didn't implement them.

Removed game.status, as it is redundant and game.winner provides more
information.

game.score will now be kept track of in the main scope, not as a part of
game, because the score is to persist across multiple games.

Removed unnecessary getters, as that is a Java convention, not a Python
convention

game.whoGoesFirst and game.run go in the main file.

game.reset has been removed. Instead, a new object is to be initialized

https://stackoverflow.com/a/1301409: Weird stuff happens when you have
methods or variables starting with __. I have replaced these with just a
single underscore, because the convention suggests using that for what
you appear to be trying to do.

Reimplemented the function to get a winner from a board state to use
looping rather than many explicit comparisons so it can be more easily
extended to a larger board.

Improved the code to generate a printout of the board so it is more
extendable, better looking, and displays keys that can be pressed

Removed ComputerPlayer and RandomPlayer from the spec, as they are not
needed and were not implemented

Renamed tic_tac_toe.py to main.py to make it more clear which is
intended to be run directly

Re-implemented control logic to handle new class structure

I have fixed a bug where if the game ends in a tie, it still assigns a
winner